api, http methods

+++ RESTful-API/inc/transacties/delete.php
@@ -1,11 +1,15 @@
 <?php
-check_required_fields(["tr_id"]);
+if (!isset($_GET['id']) || empty($_GET['id'])) {
+    die('{"error":"Missing transaction ID","status":"fail"}');
+}
+
+$tr_id = $conn->real_escape_string($_GET['id']);
 
 if(!$stmt = $conn->prepare("delete from Transacties where tr_id = ?")){
 	die('{"error":"Prepared Statement failed on prepare","errNo":' . json_encode($conn -> errno) .',"mysqlError":' . json_encode($conn -> error) .',"status":"fail"}');
 }
 
-if(!$stmt -> bind_param("i", $postvars['tr_id'])){
+if(!$stmt -> bind_param("i", $tr_id)){
 	die('{"error":"Prepared Statement bind failed on bind","errNo":' . json_encode($conn -> errno) .',"mysqlError":' . json_encode($conn -> error) .',"status":"fail"}');
 }
 $stmt -> execute();
